trying to fix the mystical bug
If prev is not null and no match if found, always encode both remaining
chunks.
Previously some chunks of chunkSize could have been stored as TempChunks
in the recipe instead of as StoredChunks with hashes and an Id.

This did not fix the mystical bug.
But it helped finding where it came from.